@@ -9,7 +9,7 @@ class HuaweiInAppUpdate {
   static const MethodChannel _channel =
       const MethodChannel('huawei_in_app_update');
 
-  static UpgradeInfo _upgradeInfo;
+  static UpgradeInfo _upgradeInfo = UpgradeInfo(false);
 
   /// Initiate to check for update from Huawei Server
   /// and retrieve App Update information
@@ -65,7 +65,7 @@ class HuaweiInAppUpdate {
   /// If force is true, dialog will not dismissible
   /// with back button or background.
   static void showUpdateDialog({
-    @required BuildContext context,
+    required BuildContext context,
     bool force = false,
     bool useRootNavigator = false,
   }) async {
@@ -86,80 +86,80 @@ class UpgradeInfo {
   final bool updateAvailable;
 
   /// App ID
-  final String appId;
+  final String? appId;
 
   /// App name
-  final String appName;
+  final String? appName;
 
   /// Package name parsed from the APK.
-  final String packageName;
+  final String? packageName;
 
   /// Size of the differential package.
-  final int diffSize;
+  final int? diffSize;
 
-  final String diffDownUrl;
+  final String? diffDownUrl;
 
-  final String diffSha2;
+  final String? diffSha2;
 
   /// Indicates whether the signature is different.
   ///
   /// The options are as follows:
   /// - **0**: no
   /// - **1**: yes
-  final int sameS;
+  final int? sameS;
 
   /// Size of the app of the target version.
-  final int size;
+  final int? size;
 
   /// App release time, accurate to day. The value is in the yyyy-MM-dd format.
-  final String releaseDate;
+  final String? releaseDate;
 
   /// URL of the app icon
-  final String icon;
+  final String? icon;
 
   /// Internal version number parsed from the APK.
-  final int versionCode;
+  final int? versionCode;
 
   /// Target version name.
-  final String versionName;
+  final String? versionName;
 
   /// Source version number.
-  final int oldVersionCode;
+  final int? oldVersionCode;
 
   /// Source version name.
-  final String oldVersionName;
+  final String? oldVersionName;
 
   /// URL for downloading the complete or differential app package.
-  final String downUrl;
+  final String? downUrl;
 
   /// New features
-  final String newFeatures;
+  final String? newFeatures;
 
   /// App update time description, for example, "released on X."
   /// The value is calculated by the server and displayed on the client.
-  final String releaseDateDesc;
+  final String? releaseDateDesc;
 
   /// ID of the AppGallery request details API.
-  final String detailId;
+  final String? detailId;
 
-  final String fullDownUrl;
+  final String? fullDownUrl;
 
-  final int bundleSize;
+  final int? bundleSize;
 
   /// Indicates whether the app is a Huawei-developed app.
   ///
   /// The options are as follows:
   /// - **0**: no
   /// - **1**: yes
-  final int devType;
+  final int? devType;
 
   /// Indicates whether the update is forcible.
-  final bool isCompulsoryUpdate;
+  final bool? isCompulsoryUpdate;
 
-  final bool isAutoUpdate;
+  final bool? isAutoUpdate;
 
   /// Reason for not recommending the update.
-  final int notRcmReason;
+  final int? notRcmReason;
 
   UpgradeInfo(
     this.updateAvailable, {
@@ -228,9 +228,9 @@ class UpdateDialog extends StatelessWidget {
   final bool forceUpdate;
 
   UpdateDialog({
-    @required this.info,
+    required this.info,
     this.forceUpdate = false,
-  }) : assert(info != null);
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -273,15 +273,15 @@ class UpdateDialog extends StatelessWidget {
                           const SizedBox(height: 20.0),
                           _buildTextRow(
                             'App',
-                            info?.appName,
+                            info.appName,
                           ),
                           _buildTextRow(
                             'Version',
-                            info?.versionName,
+                            info.versionName,
                           ),
                           _buildTextRow(
                             'Size',
-                            (info?.bundleSize ?? info?.size)?.formatBytes(),
+                            (info.bundleSize ?? info.size)?.formatBytes(),
                           ),
                           const SizedBox(height: 20.0),
                           Text(
@@ -291,7 +291,7 @@ class UpdateDialog extends StatelessWidget {
                             ),
                           ),
                           Text(
-                            info?.newFeatures,
+                            info.newFeatures ?? '',
                             style: TextStyle(
                               fontSize: 14.0,
                             ),
@@ -351,9 +351,9 @@ class UpdateDialog extends StatelessWidget {
         child: TextButton(
           style: buttonStyle,
           onPressed: () async {
-            if (info?.appId != null && info?.packageName != null) {
+            if (info.appId != null && info.packageName != null) {
               final appGalleryDeeplink =
-                  'appmarket://details?id=${info?.packageName}';
+                  'appmarket://details?id=${info.packageName}';
               if (await canLaunch(appGalleryDeeplink)) {
                 await launch(appGalleryDeeplink);
               } else {
@@ -385,18 +385,18 @@ class UpdateDialog extends StatelessWidget {
     );
   }
 
-  Widget _buildTextRow(String title, String description) {
+  Widget _buildTextRow(String? title, String? description) {
     return Row(
       children: [
         Text(
-          title,
+          title ?? '',
           style: TextStyle(
             fontSize: 14.0,
           ),
         ),
         const SizedBox(width: 8.0),
         Text(
-          description,
+          description ?? '',
           style: TextStyle(
             fontSize: 14.0,
             fontWeight: FontWeight.bold,
